requirements.txt
streamlit>=1.41.0,<2.0.0
python-docx>=1.2.0,<2.0.0
